Distribute a pkg-config .pc file
The pkg-config tool allows projects to query which packages
are installed on the target system in a distro-agnostic way.
It does so by reading a .pc file that is installed by each
individual package.

JuliusLib, which is the shared module of Julius, is intended
to be distributed and used, along the provided headers, through
a statically-linked library. It doesn't, however, install a .pc
file so applications can check whether JuliusLib is installed
or not.

This patch adds julius.pc.in file to the project tree, and adapts
the build files to generate a .pc file and install it. The generated
pkg-config file is distributed according to the major version,
and is filled in during compile-time.