libsent: unconditionally include avx target
Libsent uses avx features, but is not sets the
avx2 target, which causes various compiler errors
and breaks the building of libsent.

The compile-time errors are caused because the
configure script only adds '-mavx' target to CFLAGS
when the test import goes wrong. On standard systems,
however, the test import runs ok but we still need
to set the '-mavx' target.

Fix that by always adding avx target to CFLAGS.

[julius-speech/julius#37]