libjulius: add --enable-visualize configure flag
When compiling Julius with Gtk+ support for visualization, it
needs some structures from libjulius that are conditionally
compiled based on VISUALIZE define.

To try and define it at compile time, Julius uses --enable-visualize
flag. Libjulius, however, doesn't use that flag and ends up
not defining this variable, which causes compilation errors
with --enable-visualize.

Fix that by adding a configure flag for --enable-visualize.