visual: show word list in a sidebar
For the sake of sanity, the drawing code omits
some words from being rendered, otherwise it'd
be too hard to understand the backtrellis.

The problem with that approach is that we don't
know where some branchs point to, since some words
are not rendered.

Fix that by showing a sidebar with the words, in
the order they're displayed in the graph.